+++ docs/changelog.md
@@ -18,6 +18,7 @@ Click on the version in the following list to see all changes since the previous
 * test: remove todo for deployment with licensing disabled after successful test
 * feat: health check for Scrum Poker app available in configuration (#67)
 * feat: add check for validity of license to health check (#67)
+* test: implement test coverage for health check logic
 
 ### [4.0.1] - 2019-06-16